# repo/docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/alerts": {
      "get": {
        "operationId": "AlertsController_findAll",
        "parameters": [
          {
            "name": "includeTriggered",
            "required": false,
            "in": "query",
            "description": "Include already-triggered alerts",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Array of alerts"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List price alerts",
        "tags": [
          "Alerts"
        ]
      },
      "post": {
        "operationId": "AlertsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAlertDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Alert created"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create a new price alert",
        "tags": [
          "Alerts"
        ]
      }
    },
    "/api/alerts/{id}": {
      "delete": {
        "operationId": "AlertsController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Alert UUID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Deleted"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Delete a price alert",
        "tags": [
          "Alerts"
        ]
      }
    },
    "/api/push/vapid-public-key": {
      "get": {
        "operationId": "PushController_getVapidPublicKey",
        "parameters": [],
        "responses": {
          "200": {
            "description": "VAPID public key"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get VAPID public key for push subscription",
        "tags": [
          "Push"
        ]
      }
    },
    "/api/push/subscribe": {
      "post": {
        "operationId": "PushController_subscribe",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SubscribePushDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Subscription saved"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Register a push subscription",
        "tags": [
          "Push"
        ]
      },
      "delete": {
        "operationId": "PushController_unsubscribe",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UnsubscribePushDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Subscription removed"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Remove a push subscription",
        "tags": [
          "Push"
        ]
      }
    },
...
  },
  "info": {
    "title": "Crypto Dashboard API",
    "description": "Final project for BMU1208 Web-Based Programming",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      },
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "refresh_token"
      }
    },
    "schemas": {
...
      "CreateAlertDto": {
        "type": "object",
        "properties": {}
      },
      "SubscribePushDto": {
        "type": "object",
        "properties": {}
      },
      "UnsubscribePushDto": {
        "type": "object",
        "properties": {}
      }
    }
  }
}